bootconfig is part of VM identity
VM's identity should be changed if debug level is changed. Debug level
is implemented using bootconfig. So changing the bootconfig should
invalidate the previous identity. This change implements the policy by
adding the bootconfig to MicrodroidData and refuses to boot if the saved
bootconfig is not the same as the current bootconfig.

Bug: 208639280
Test: atest MicrodroidTestApp on oriole_pkvm
Change-Id: I63ac0c2ecd7514d9193e655c9258c305c87e2a08