Add support for STANDALONE_SYSTEMSERVER_JARS
Stop differentiating between the different ways an APEX can contribute
to derive_classpath - it makes no difference to us, and it doesn't
scale if new environment variables are added.

Remove the workaround to stop the tests failing. Extended the existing
temporary workaround for setting the environment variables in the VM.

Bug: 210472252
Test: atest ApexTestCases
Test: atest ComposTestCase
Test: composd_cmd staged-apex-compile
Change-Id: I97b48cf71c6f22ade27c1ca2e0bad1bb9ab582fc