Implement the DICE HAL using fake values
THe values are placeholders until true values are made available for the
HAL to pick up. Adding the HAL allows diced to connect to it and get
some values rather than it trying to connect to the HAL for a little
bit, failing and then generating its own fake values. The result is
still fake, but there it less waiting and the components are connecting
as they are expected to.

Bug: 214231981
Test: run microdroid
Change-Id: I3e6d2995e388418f99d1987fa1af22f9818bd745