Use protected VM if possible
If the system is capable of running a protected VM, then request one.

On userdebug/eng builds only, fall back to unprotected VM if that is
all that is available. (This is useful for testing.)

A protected VM doesn't support full debug, so use app only debug in
that case. That change requires relaxing the constraint on running
compilation in test mode to match.

I've also increases the VM start timeout to avoid spurious timeouts.

Bug: 217687661
Test: composd_cmd staged-apex-compile
Test: composd_cmd test-compile
Change-Id: Ic7be2c5260982c3c69421c01d10c1537864b87be