Disable kmsg when microdroid isn't fully debuggable
The default behaviour is to rate limit logs from user space to kmsg so
override that to disable the logging instead. This prevents logs from
the microdroid system components from showing in the kernel logs while
the kernel log is still output on the uart.

Make it so microdroid_managed doesn't crash if it can't use /dev/kmsg.

Bug: 219743539
Test: MicrodroidTests
Change-Id: I68c8335176b088f2a65b6effa77d1cfb9adf1939