Use AtomicU64 for handle_ref_count
This allows us to use a read lock instead of a write on `lookup`.

Intriguingly, the corresponding call to decrease the count in `forget`
does not need a lock, thanks to Rust. So this should be a net
performance gain.

Also, panic is a handle's reference count is about to overflow.

Bug: 220386264
Test: atest AuthFsHostTest
Change-Id: Icc3ba1a3fe92431837d6cffb470688ae0883a355