Synchronize VM callbacks
Both the callback and the executor are needed to invoke a callback so
read and update them atomically. This avoids the possibility of the
callback being called on the wrong executor.

Also make sure that onDied() is only called once for a VM by
synchronizing the possible death sources.

Fix: 220730550
Test: atest MicrodroidTests
Change-Id: I434ea70cc90f1df9a23c8518798ef26fe0adf9e7