Add hooks for IsolatedCompilation metrics
Enhance the CompOS service to report more details on failure.

Create a new IsolatedCompilationMetrics class to handle reporting when
compilation is scheduled and when it succeeds/fails.

No metrics are logged yet - we first need to define the proto format
in atoms.proto.

Additionally, improve the handling of multiple calls to onCompletion so
only the first one does anything. (We can get binderDied() called
after success, since composd exits.)

Bug: 218525257
Test: stage an APEX, run the job, observe logs
Test: run composd_cmd, cause it to fail in various ways
Change-Id: I4f97a82ddb5c1d53d67921e1e7ecc6c61229fb76