zipfuse: Support mounting as noexec
Add an option to mount while disallowing the execution of binary files.
This allows for a layer of defence when mounting something that is only
expected to contain data. Although the data may still be treated as a
script, the interpreter for that script will have to come from another
source.

The `zipfuse` command takes the `--noexec` flag to enable this option.

Bug: 237742605
Test: atest ZipFuseTest
Change-Id: I732c96109ceaf65716c2c3ac27862c65bb7bd81f