vsock_test: Wait for ACK from server
vsock_test has been flaky - the test in host would receive ECONNRESET
when reading from the client's vsock socket. This was due to the guest
exiting immediately after writing to the socket. This in turn terminated
crosvm, including the vhost-vsock backend, and reset the socket.

Fix this by having the guest wait for a reply from the host before
shutting down and explicitly shutting down the connection.

Bug: 238733740
Test: atest VirtualizationTestCases.64
Change-Id: I36c678f469fc3269902f5d8b1eb73cd431661ae6