Make authfs start optional
Only start the service if the VM config requests it.

Make CompOS explicitly request authfs, since it requires it.

Also improve the error messages if authfs is missing (which I
triggered while working on this).

This doesn't include restricting the use of the config option, but we
have a seperate bug to handle that in general.

Bug: 245262525
Test: atest MicrodroidTests MicrodroidHostTestCases
Change-Id: I4363daf0f5cfb0c1c7ffbb7ae2ca2b9cc395167d