pvmfw: Use pvmfw_embedded_key from Rust
pvmfw will need the key to validate AVB-signed images so wrap it in a
rlib to import it into Rust code while making sure Soong is aware of the
dependency and handles file paths.

As this will be used in the context of AVB, create the module and
re-export the PUBLIC_KEY from there.

Use a debug! statement in main() to ensure that the key is kept in the
final binary by the toolchain.

Bug: 255575669
Test: m pvmfw_bin # verify that it contains testkey_rsa4096_pub.bin
Change-Id: I840a32a39f113d6becd50bea42b8552e551fee5f