Compos test compilation is only for dogfooders and is critical
For dogfooders (to be specific, userdebug/eng builds), compos test
compilation is triggered daily (when the device is idle of course) even
without ART mainline updates. If it fails to start the VM, composd is
killed - instead of reporting it back to system_server where it is
silently ignored. The composd crash report will then automatically be
written to the dropbox so that dogfooders have a chance to upload it.

Bug: 250929504
Test: `adb shell cmd jobscheduler run -f android 5132250`
and `adb shell "kill -11 $(pidof composd)"`

Change-Id: Ia9bb1597f8f9ead7b88dcce2e92adf61bdcafc7e