Cleanup Android.bp:framework-virtualization
There are a bunch of variables that are already set in the
framework-module-defaults, so there is no need to duplicate them.

Also disable the JavaApiUsedByMainlineModule check, since we are not an
updatable module we don't need it, and it's quite spammy.

As a side effect of this change our tests & demo apks need to compile
against framework-virtualization.impl library (which includes the hidden
APIs). Once the APIs are promoted to @SystemApi we can change the tests
to compile against framework-virtualization (which contains the API
stubs).

Bug: 243512044
Test: presubmit
Change-Id: Ib614d266f7a9cb8517afccd4feb737068b51fb20