Make sure a deleted VM stays deleted
If we delete a VM instance it is removed from the weak reference
map. If we then recreate a VM with the same name, we end up with two
VirtualMachine objects, which will inevitably end badly. So make sure
that deletion is sticky.

Added a test to check this (and made sure it correctly failed
initially - which also revealed it wasn't cleaning up properly on
failure).

Bug: 261037705
Test: atest MicrodroidTests
Change-Id: I852d5547a24c770e4b97252be5056e06fe3511af