Add --uid --gid options to zipfuse
So far, files and directories in a zipfuse filesystem had root as its
uid/gid. This change adds two options --uid and --gid to zipfuse which
are used to override the behavior.

For now, microdroid_manager uses `--uid 0 --gid 0` which means there's
no behavioral change. A follow-up change will set the IDs to a non-root
value.

Bug: 264668376
Test: run microdroid tests
Change-Id: I935424a52edcc3868fdf6aef526a641855a67f63