virtmgr: Delegate debugVmList to virtualizationservice
The shell user has permission to list all VMs in the system (called from
'vm list'). This stopped working after virtmgr was introduced, as it
only has visibility of its own VMs. Add a new method on
VirtuliazationService_Internal to retrieve this information from
the global VS, which does keep a list of the allocated VM contexts.

Extend VS's info with the requestor's UID and PID to match the original
functionality. The one piece of information missing is the VM's running
state.

Bug: 245727626
Bug: 264826962
Test: adb shell /apex/com.android.virt/vm list
Test: vm/vm_shell.sh start-microdroid --auto-connect
Test: atest MicrodroidHostTestCases
Change-Id: I4cf5d1a9792aca995b63a13c2ee47c6033d9a5d6