Remove support for --daemonize
The 'daemonize' feature allowed a VM spawned by shell to be held by
VirtualizationService indefinitely, eg. if adb was disconnected.
The setup is not compatible with the new virtmgr/virtualizationservice
split and can be replaced with 'adb shell vm run ... &' anyway. All
tests and scripts have been migrated away from it, so we can remove it
now.

Bug: 265248057
Bug: 245727626
Test: atest -p packages/modules/Virtualization:avf-presubmit
Change-Id: Iba4fc40ffd9ddc59afaa535cf746872b1d4ba87a