Test that we can parcel & unparcel VirtualMachineDescriptor
The test creates a VirtualMachineDescriptor, writes it to a parcel,
reads it from the parcel, and then tries to create a VM from it.

This change also fixes the bug in how VirtualMachineDescriptor is
parceled/unparceled: the writeToParcel was writing a full
ParcelFileDescriptor parcelable, while VirtualMachineDescriptor(Parcel)
c-tor was reading just a raw fd.

Bug: 259384440
Test: atest MicrodroidTestApp
Change-Id: I809fa635ebb934e0ba957d3e2642d1b40d21ec50