pvmfw: Log clarification about initrd_normal error
As pvmfw_avb::verify_payload() checks if the payload is debuggable by
first trying to verify a payload named 'initrd_normal' then an
'initrd_debug', debuggable payloads will fail to verify the former while
still being valid. As libavb logs any verification failure, this will
result in the confusing log:

    avb_slot_verify.c:443: ERROR: initrd_normal: Hash of data does not match digest in descriptor.

To avoid any confusion, detect a successfully verified payload and log a
message to clarify the above in pvmfw. This is prefered over logging at
the libavb call site so that we don't introduce logging in pvmfw_avb.

Bug: 284443878
Bug: 289173497
Test: TH
Change-Id: Icca19ebcb021e4280b6b80715045b0239adb52ab