Fix race between read/write
wait_for_file waits for "creation". The test is flaky due to the race
between read/write.

Added a WARNING note to the function and fixed the test to avoid the
race.

Verified that the current usages are okay since they wait for files
which are made visible by mounting or similar, not writing from other
processes/threads.

Bug: 292297891
Test: atest microdroid_manager_test
Change-Id: I05e99aa433fe7664c2d74ca1f4f34bc20bfd1939