Simplify ziputils
zip_sections was taking a reader by value and then returning it.

But we don't have to do that - for both Read and Seek, if X satisfies
the trait then so does &mut X. So if the caller wants to retain their
reader, they can just pass us a mutable reference to it.

Bug: 299591171
Test: atest libapkverify.test libapkverify.integration_test
Change-Id: Ic849d3c15843a712c8014c342569a42e7becf4b8