Refactoring: split Microdroid Manager main.rs
It's way too big, and it's making it hard for me to see where to make
my next change.

Move APK / APEX verification to verify.rs, and DICE derivation to
dice.rs.

Rename what's left of the old dice.rs to dice_driver.rs, since its
role is now just to communicate with the driver.

And a few other minor tweaks. No behavior change is intended.

Bug: 299591171
Test: atest MicrodroidTests
Change-Id: Ibb25473d4d969d64a782babfa7a5f4a7874c1cdf