Refactor Dice policy test
A substantial portion of unit testing dice policy structure is about
creating test dice_chains, constraint_spec & expected dice policy.

TestArtifacts is a struct to encapsulate these, along with that we
provide method to get an example TestArtifacts, which is used by the
tests.

As we will see, it will be useful as we need more dice chains for
comprehensive testing.

Test: atest libdice_policy.test
Bug: 291238565
Change-Id: Ie9b96e231ded334cc2cb370aa8a320a8a5e9c757